Introduce LucraClient v1 singleton with event-based API
- Replace constructor-based onMessage handlers with .on()/.off() event listener API backed by EventTarget
- Enforce singleton pattern via LucraClient.initialize(), getInstance(), and destroy()
- Remove LucraClientBase from public exports; LucraClient is now the sole public entry point
- Fix origin validation on incoming postMessages to prevent spoofed events from other windows
- Fix deep link URL validation to use origin comparison instead of string prefix match
- Fix duplicate listener leak when same function registered for multiple event types
- Fix iframe cleanup in close() to use element removal instead of iinnerHTML clear
- Add null/type guard on event.data to prevent crash on malformed postMessages
- Add try/catch around matchupInviteUrl transformer to handle rejections
- Fix tag-and-release workflow to require merged == true before triggering a release
- Add CI workflow to run tests and build on PRs and pushes to main
- Update all docs to reflect v1 API, remove v0 references
