+++ .github/workflows/tag-and-release.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   release:
     runs-on: ubuntu-latest
-    if: startsWith(github.event.pull_request.title, 'Release:')
+    if: startsWith(github.event.pull_request.title, 'Release:') && github.event.pull_request.merged == true
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
@@ -22,6 +22,8 @@ jobs:
           node-version: "22.x"
       - name: Install dependencies
         run: bun install
+      - name: Test
+        run: bun test
       - name: Build
         run: bun run build
       - name: Set up Git
