+++ .gitignore
@@ -172,3 +172,5 @@ out
 
 # Finder (MacOS) folder config
 .DS_Store
+
+.projects
