+++ docs/0.0_README.md
@@ -7,6 +7,8 @@
 ## Initialization
 [LucraClient Initialization](1.2_initialize_client.md)
 
+Import from `lucra-web-sdk` for the singleton API (`initialize` / `getInstance` / `destroy`).
+
 ## Lucra Flows
 [Lucra Flows](1.3_lucraflows.md)
 
@@ -18,3 +20,12 @@
 
 ## Lucra Event Listener
 [Lucra Event Listener](1.6_lucra_event_listener.md)
+
+## Convert to Credit
+[Convert to Credit](1.7_convert_to_credit.md)
+
+## Updating Users
+[Updating Users](1.8_updating_users.md)
+
+## Rewards
+[Rewards](1.9_rewards.md)
