+++ docs/1.2_initialize_client.md
@@ -1,46 +1,56 @@
 # Initializing the LucraClient
 
-### Creating a Singleton (Recommended for Frontend Libraries)
+Import from `lucra-web-sdk`. The client enforces a singleton - only one instance can exist at a time.
 
-When using the LucraClient in frontend frameworks like React, Vue, or Angular, it's important to create a singleton instance outside of component render cycles. Creating the client inside a component body or render function will cause a new instance to be created on every render, which can lead to memory leaks and unexpected behavior.
-
-**Recommended approach:**
+**Initialize once at app startup:**
 
 ```typescript
-// lucraClient.ts (or similar module file)
 import { LucraClient } from 'lucra-web-sdk';
 
-let lucraClientInstance: LucraClient | null = null;
-
-export function getLucraClient(): LucraClient {
-  if (!lucraClientInstance) {
-    lucraClientInstance = new LucraClient({
-      tenantId: "<your Lucra tenant id>",
-      env: "sandbox" | "production",
-      onMessage: {
-        // callback functions
-      },
-    });
-  }
-  return lucraClientInstance;
-}
+LucraClient.initialize({
+  apiKey: "<your Lucra api key>",
+  tenantId: "<your Lucra tenant id>",
+  env: "sandbox",
+});
 ```
 
-**In your component:**
+**Retrieve the instance anywhere:**
 
 ```typescript
-import { getLucraClient } from './lucraClient';
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+```
+
+Calling `initialize` a second time throws. Calling `getInstance` before `initialize` throws.
+
+**Cleanup (e.g. on logout):**
+
+```typescript
+LucraClient.destroy();
+```
+
+After `destroy`, `initialize` can be called again.
+
+**React example:**
+
+```typescript
+// app entry point (e.g. main.tsx)
+import { LucraClient } from 'lucra-web-sdk';
+
+LucraClient.initialize({
+  apiKey: "<your Lucra api key>",
+  tenantId: "<your Lucra tenant id>",
+  env: "production",
+});
+```
+
+```typescript
+// any component
+import { LucraClient } from 'lucra-web-sdk';
 
 function MyComponent() {
-  // ✅ Good: Use the singleton instance
-  const lucraClient = getLucraClient();
-  
-  // ❌ Bad: Don't create inside component body
-  // const lucraClient = new LucraClient({ ... });
-  
-  // ❌ Bad: Don't create inside useEffect without proper cleanup
-  // useEffect(() => {
-  //   const client = new LucraClient({ ... });
-  // }, []);
+  const client = LucraClient.getInstance();
+  // ...
 }
 ```
