+++ docs/1.3_lucraflows.md
@@ -2,34 +2,38 @@
 
 ### Open Lucra
 
-```
-const lucraClient = new LucraClient(...)
-lucraClient.open(element: <HTMLElement that will contain the iframe for Lucra>, phoneNumber?: string)
+```typescript
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+client.open(element, phoneNumber?)
   // and one of the below
   .profile()
-  .home(locationId?: string)
+  .home(locationId?)
   .deposit()
   .withdraw()
-  .createMatchup(gameId?: string)
-  .matchupDetails(matchupId: string)
-  .tournamentDetails(matchupId: string)
-  .deepLink(url: string)
+  .createMatchup(gameId?)
+  .matchupDetails(matchupId)
+  .tournamentDetails(matchupId)
+  .deepLink(url)
 ```
 
 Where phone number (if valid) will automatically send the SMS verification code, skipping the step where the user must enter phone number manually.
 
 ### Redirect Lucra
 
-```
-// use an existing LucraClient that's open
-lucraClient.redirect()
+```typescript
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+client.redirect()
   // and one of the below
   .profile()
-  .home(locationId?: string)
+  .home(locationId?)
   .deposit()
   .withdraw()
-  .createMatchup(gameId?: string)
-  .matchupDetails(matchupId: string)
-  .tournamentDetails(matchupId: string)
-  .deepLink(url: string)
+  .createMatchup(gameId?)
+  .matchupDetails(matchupId)
+  .tournamentDetails(matchupId)
+  .deepLink(url)
 ```
