+++ docs/1.4_deeplinks.md
@@ -2,33 +2,33 @@
 
 #### lucraClient.sendMessage.deepLinkResponse(...)
 
-Responds to a `deepLink` request message with a transformed URL. Used when Lucra requests a URL that should be opened in your application, which then navigates back to Lucra at the specified location. For matchup invitation links, prefer using `matchupDeepLinkHandler` instead.
+Responds to a `deepLink` event with a transformed URL. Used when Lucra requests a URL that should be opened in your application, which then navigates back to Lucra at the specified location. For matchup invitation links, prefer using `setMatchupDeepLinkHandler` instead.
 
-**Example payload:**
+**Example:**
 ```typescript
-lucraClient.sendMessage.deepLinkResponse({
-  url: "https://yourapp.com?redirect=https://tenant.lucrasports.com/app/matchups/123"
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+
+client.on('deepLink', ({ url }) => {
+  client.sendMessage.deepLinkResponse({
+    url: `https://yourapp.com?redirect=${url}`
+  });
 });
 ```
 
-
 ### Configuring Matchup Invite URL Generation
 
-When users want to share or invite others to a matchup, you need to configure how invite URLs are generated. Use the `matchupDeepLinkHandler`:
+When users want to share or invite others to a matchup, configure how invite URLs are generated using `setMatchupDeepLinkHandler`:
 
 ```typescript
-const lucraClient = new LucraClient({
-  tenantId: "your-tenant-id",
-  env: "production",
-  onMessage: {
-    /* ... */
-  },
-});
+import { LucraClient } from 'lucra-web-sdk';
 
-// Configure the transformer to generate invite URLs
-lucraClient.matchupDeepLinkHandler = (matchupId) => {
+const client = LucraClient.getInstance();
+
+client.setMatchupDeepLinkHandler((matchupId) => {
   return Promise.resolve(`${window.location.origin}?matchupId=${matchupId}`);
-};
+});
 ```
 
 **How it works:**
@@ -39,43 +39,10 @@ lucraClient.matchupDeepLinkHandler = (matchupId) => {
 4. **Your app detects the matchupId** - On launch, check URL parameters and navigate to the matchup:
 
 ```typescript
-// On app initialization, check for invite parameters
 const urlParams = new URLSearchParams(window.location.search);
-const matchupId = urlParams.get("matchupId");
+const matchupId = urlParams.get('matchupId');
 
 if (matchupId) {
-  // Open Lucra to the home screen first then navigate to matchup details
-  lucraClient.open(containerElement).matchupDetails(matchupId);
+  client.open(containerElement).matchupDetails(matchupId);
 }
-```
-
-#### Migration from deprecated `deepLinkHandler` for matchup invitations
-
-**Key Differences:**
-
-| Old (`deepLinkHandler`)        | New (`matchupDeepLinkHandler`)       |
-| ------------------------------ | ------------------------------------ |
-| Received full Lucra URL        | Receives clean `matchupId` string    |
-| Required manual URL parsing    | No parsing needed                    |
-| Manual `deepLinkResponse` call | Returns URL directly (auto-responds) |
-| Generic for all deep links     | Specific to matchup invites          |
-
-**Migration Example:**
-
-```typescript
-// OLD APPROACH
-lucraClient.deepLinkHandler = ({ url }) => {
-  // Had to parse the Lucra URL to extract matchupId
-  const lucraUrl = new URL(url);
-  const matchupId = lucraUrl.pathname.split("/").pop();
-
-  const shareUrl = `${window.location.origin}?matchupId=${matchupId}`;
-  lucraClient.sendMessage.deepLinkResponse({ url: shareUrl });
-};
-
-// NEW APPROACH (Recommended)
-lucraClient.matchupDeepLinkHandler = (matchupId) => {
-  // Receive matchupId directly, just return the URL
-  return Promise.resolve(`${window.location.origin}?matchupId=${matchupId}`);
-};
 ```
