+++ docs/1.5_exiting_lucra.md
@@ -1,15 +1,25 @@
-### LucraClient - Listening for "exit Lucra" 
+# Listening for "Exit Lucra"
 
-### Registering `exitLucraHandler` function
-
-To determine when a user has attempted to exit the Lucra experience, you may register a handler function after successful login like so:
+Register a listener for when a user attempts to exit the Lucra experience:
 
 ```typescript
-// Register the exit lucra handler after we are subscribed
-lucraClient.exitLucraHandler = () => {
-  console.log("User attempted to exit Lucra!");
-};
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+
+client.on('exitLucra', () => {
+  console.log('User attempted to exit Lucra!');
+});
 ```
 
-An example of where to register this would be in the `loginSuccessHandler` (or after the login event has fired) to ensure the user's Lucra experience is fully ready.
+A good place to register this is inside your `loginSuccess` listener, once the user's Lucra session is fully ready.
 
+To stop listening:
+
+```typescript
+const handleExit = () => { /* ... */ };
+
+client.on('exitLucra', handleExit);
+// later...
+client.off('exitLucra', handleExit);
+```
