+++ docs/1.6_lucra_event_listener.md
@@ -1,31 +1,61 @@
 # Lucra Event Listener
 
-### Registering `onMessage` functions
+## v1: .on() / .off()
 
-If you need to register a handler function for any of the `onMessage` callbacks, you can do so after instantiation like so:
+Subscribe to events from Lucra using `.on()`. The listener receives a typed `data` payload specific to each event.
 
 ```typescript
-const lucraClient = new LucraClient(...)
-lucraClient.userInfoHandler = handleUserInfo;
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+
+client.on('loginSuccess', (user) => {
+  console.log('User logged in', user);
+});
+
+client.on('matchupCreated', ({ matchupId }) => {
+  console.log('Matchup created:', matchupId);
+});
 ```
 
-### Messages from LucraClient
+To stop listening, pass the same function reference to `.off()`:
 
-- `loginSuccess` - the user has logged into Lucra
-- `userInfo` - whenever an update happens to the user, the callback to this function will receive the newest version of that user
-- `matchupCreated` - the user successfully created a matchup, and contains the id of that matchup
-- `matchupStarted` - the matchup owner started the matchup, and contains the id of that matchup
-- `matchupAccepted` - the user successfully joined someone else's matchup, and contains the id of that matchup
-- `matchupCanceled` - the user successfully canceled the matchup, and contains the id of that matchup
-- `matchupInviteUrl` - Lucra is requesting an invite URL for a specific matchup. Use `matchupDeepLinkHandler` to handle this. Your handler receives the `matchupId` and must call `sendMessage.deepLinkResponse()` with the generated URL.
-- `convertToCredit` - if Convert to Credit is enabled, Lucra will call this method with the desired amount to convert to credit. Respond with a `convertToCreditResponse` message defined below
-- `deepLink` - Lucra is requesting a url that the user of the SDK will open to then open up the LucraClient at where the deepLink url is linking to. For matchup invitations, use `matchupDeepLinkHandler`.
-- `tournamentJoined` - the user successfully joined a tournament, and contains the id of the tournament
-- `navigationEvent` - the user has navigated somewhere in Lucra, contains the full url that you can use later when doing `.open().deepLink(url)`
-- `claimReward` - the user is redeeming a Free To Play reward
+```typescript
+const handleUserInfo = (user) => { /* ... */ };
+
+client.on('userInfo', handleUserInfo);
+// later...
+client.off('userInfo', handleUserInfo);
+```
+
+### Available events
+
+| Event | Payload | Description |
+| --- | --- | --- |
+| `loginSuccess` | `LucraLoginSuccessBody` | User logged into Lucra |
+| `userInfo` | `LucraUserInfoBody` | User data updated |
+| `matchupCreated` | `{ matchupId: string }` | User created a matchup |
+| `matchupStarted` | `{ matchupId: string }` | Matchup owner started the matchup |
+| `matchupAccepted` | `{ matchupId: string }` | User joined someone else's matchup |
+| `matchupCanceled` | `{ matchupId: string }` | User canceled a matchup |
+| `activeMatchupStarted` | `{ matchupId: string }` | An active matchup started |
+| `tournamentJoined` | `{ matchupId: string }` | User joined a tournament |
+| `convertToCredit` | `{ amount: number }` | User requested a credit withdrawal |
+| `deepLink` | `{ url: string }` | Lucra is requesting a URL to open in your app |
+| `navigationEvent` | `{ url: string, page?: LucraPage }` | User navigated within Lucra |
+| `claimReward` | `{ reward: LucraReward }` | User is redeeming a Free To Play reward |
+| `exitLucra` | `void` | User attempted to exit Lucra |
+
+For matchup invite URL generation, see [Deeplinks](1.4_deeplinks.md).
 
 ### Messages you can send to LucraClient
 
-`userUpdated` - update the LucraClient user with known information about the current user. This information will be persisted to LucraSports and will be used to pre-fill forms the user will have to fill out for things like Identity Verification
+`userUpdated` - update the LucraClient user with known information about the current user. This information will be persisted to LucraSports and will be used to pre-fill forms the user will have to fill out for things like Identity Verification.
 
-You may optionally include a `metadata` property as valid JSON related to the Lucra user. Key/value pairs should be of type `String` only. _Note_: On the `metadata` object, the `external_id` property will be used to index, if provided. Passing in metadata in this way will be used to link client users with the Lucra user for autosettlement identification purposes.
+You may optionally include a `metadata` property as valid JSON related to the Lucra user. Key/value pairs should be of type `String` only. _Note_: On the `metadata` object, the `externalId` property will be used to index, if provided. Passing in metadata in this way will be used to link client users with the Lucra user for autosettlement identification purposes.
+
+`navigate` - navigate the Lucra iframe to a specific pathname within the Lucra app.
+
+```typescript
+client.sendMessage.navigate({ pathname: '/app/home' });
+```
