+++ docs/1.7_convert_to_credit.md
@@ -12,30 +12,29 @@ Enables the convert to credit option in the withdraw funds screen. Call this whe
 lucraClient.sendMessage.enableConvertToCredit();
 ```
 
-#### lucraClient.sendMessage.convertToCreditResponse(...)
+#### Handling the convertToCredit event
 
-Call this after receiving a `convertToCredit` message when the user attempts to withdraw funds. The user-requested amount to convert will be sent along in the original message:
-
-```
-// TypeScript
-export type LucraConvertToCreditBody = { amount: number }
-```
-
-**Example payload:**
+Listen for the `convertToCredit` event and respond with a `convertToCreditResponse` message:
 
 ```typescript
-lucraClient.sendMessage.convertToCreditResponse({
-  id: "xyz-123",
-  title: "Game Credits",
-  iconUrl: "https://example.com/icon.png",
-  conversionTerms: "Instant transfer",
-  convertedAmount: "<amount from the convertToCredit message>",
-  convertedDisplayAmount: "<desired formatted version of amount>",
-  shortDescription: "Convert to game credits",
-  longDescription:
-    "Withdraw your winnings as game credits that can be used immediately in-game.",
-  metaData: {
-    ...
-  },
+import { LucraClient } from 'lucra-web-sdk';
+
+const client = LucraClient.getInstance();
+
+client.on('convertToCredit', ({ amount }) => {
+  client.sendMessage.convertToCreditResponse({
+    id: 'xyz-123',
+    title: 'Game Credits',
+    iconUrl: 'https://example.com/icon.png',
+    conversionTerms: 'Instant transfer',
+    convertedAmount: amount,
+    convertedDisplayAmount: `$${amount.toFixed(2)}`,
+    shortDescription: 'Convert to game credits',
+    longDescription:
+      'Withdraw your winnings as game credits that can be used immediately in-game.',
+    metaData: {
+      // ...
+    },
+  });
 });
 ```
