+++ docs/1.8_updating_users.md
@@ -26,3 +26,5 @@ lucraClient.sendMessage.userUpdated({
   }
 });
 ```
+
+Note: `state` must be a valid 2-letter state code (`StateCode`). Full state names are not accepted.
