+++ tsconfig.json
@@ -27,5 +27,5 @@
     "noPropertyAccessFromIndexSignature": false
   },
   "exclude": ["dist"],
-  "files": ["index.ts", "types/types.ts"]
+  "files": ["index.ts", "types/types.ts", "base.ts", "v1.ts"]
 }
